Create database - bin/console doctrine:database:create Now, instead of running doctrine:schema:update, run:
./bin/console doctrine:migrations:diff
This created a new file in app/DoctrineMigrations.  Go open that up.

Check this out: the up() method executes the exact SQL that we would have gotten from the doctrine:schema:update command. But instead of running it, it saves it into this file. This is our chance to look at it and make sure it's perfect.

When you're ready, run the migration with:

bin/console doctrine:migrations:migrate